LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/inert-node-is-unfocusable-expected.txt

I get focus
I don't get focus.

I'm editable
I'm tabindexed.
 Link

FAIL Test that inert nodes are not focusable. assert_equals: body expected false but got true
